+++ src/daemon/rpc_command_executor.cpp
@@ -67,7 +67,7 @@ bool t_rpc_command_executor::print_transaction_pool_short() {
   return true;
 }
 
-bool t_rpc_command_executor::start_mining(cryptonote::account_public_address address, size_t num_threads) {
+bool t_rpc_command_executor::start_mining(cryptonote::account_public_address address, uint64_t num_threads) {
   std::cout << "start mining" << std::endl;
   return true;
 }
